fix: phantom parent desync — back-patch children on out-of-order arrival
Root cause: insert_relaxed() never updated a parent's children set when
the parent arrived after its child via gossip. This caused causal_order()
to produce different topological sort orders on different nodes (wrong
in-degrees for same-round blocks), diverging VRF seeds and permanently
stalling consensus after ~4000 blocks.

Fixes:
- Back-patch children in insert_relaxed() when orphaned parent resolves
- Align RoundState prune buffer to 16 rounds (was 2), matching DAG retention
- Evict stale orphan_parents during DAG prune

3 new tests, 988 total pass. Clippy clean, fmt clean.

+++ crates/aztibase-node/src/sentinel.rs
@@ -458,7 +458,7 @@ pub async fn run_sentinel(
                 gas_limit: 0,
                 empty_batches: 0,
                 total_batches_window: 0,
-                ms_since_last_finality: (stall_secs * 1000) as u64,
+                ms_since_last_finality: stall_secs * 1000,
                 peer_count: handles.peer_count.load(Ordering::Relaxed),
                 mempool_size: handles.mempool_size.load(Ordering::Relaxed),
             };
